requirements.txt
aiogram>=3.0.0      
asyncio
aiohttp>=3.8.0      
aiosqlite>=0.19.0   
beautifulsoup4>=4.12.0  
python-telegram-bot>=20.3  
python-dotenv>=1.0.0
telethon